src/beddel-py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "beddel"
version = "0.1.0"
description = "Agent-Native Workflow Engine — Declarative YAML pipelines for AI agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Botanarede", email = "dev@botanarede.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "pydantic>=2.10",
    "pyyaml>=6.0",
    "litellm>=1.55",
    "httpx>=0.28",
    "opentelemetry-api>=1.29",
    "opentelemetry-sdk>=1.29",
]

[project.optional-dependencies]
fastapi = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "sse-starlette>=2.0",
]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "pytest-timeout>=2.3",
    "pytest-mock>=3.14",
    "ruff>=0.9",
    "mypy>=1.14",
    "types-PyYAML>=6.0",
    "respx>=0.22",
]

[tool.hatch.build.targets.wheel]
packages = ["src/beddel"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
timeout = 30

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
